[fix] Refactoring main.cs
