+++ src/EasyEPlanner.csproj
@@ -204,18 +204,28 @@
       <DependentUpon>LogFrm.cs</DependentUpon>
     </Compile>
     <Compile Include="Logs\Logs.cs" />
+    <Compile Include="Main\AboutProgrammAction.cs" />
+    <Compile Include="Main\AddInModule.cs" />
+    <Compile Include="Main\BindingSynchronizationAction.cs" />
+    <Compile Include="Main\ExportTechDevsToExcelAction.cs" />
+    <Compile Include="Main\LoadDescriptionAction.cs" />
+    <Compile Include="Main\SaveAsXMLAction.cs" />
+    <Compile Include="Main\SaveDescriptionAction.cs" />
+    <Compile Include="Main\ShowDevicesAction.cs" />
+    <Compile Include="Main\ShowOperationsAction.cs" />
+    <Compile Include="Main\ShowTechObjectsAction.cs" />
     <Compile Include="ModeFrm.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="ModeFrm.Designer.cs">
       <DependentUpon>ModeFrm.cs</DependentUpon>
     </Compile>
+    <Compile Include="ModulesBindingUpdater.cs" />
     <Compile Include="PInvokeUtil.cs" />
     <Compile Include="FileSavers\PrgLuaSaver.cs" />
     <Compile Include="ProjectConfiguration.cs" />
     <Compile Include="FileSavers\ProjectDescriptionSaver.cs" />
     <Compile Include="ProjectManager.cs" />
-    <Compile Include="main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
